Working Towards Scale Detection
Created Note and Notes class for easy look up of certain notes. Note
contain their name i.e. "C#" and if their a whole or half step in the
for of a float in order to do easy math. Next task is to create a
supervised learning process to detect certain scales.