Figuring out scale detection
Currently my Note class does not contain information about the octave.
If I change it, I change the way my notes class works. Not sure if it is
nessescary to make the change. The notes class purpose is only to figure
out what the note is based on an index of the c major scale. Adding the
octave to the note will make each of the "base notes" have an octave of
zero. Thinking about creating a enum with the bit values to determine
the note rather than an array and using an index.