Scale Detection Issues
Currently the steps in between notes is off because I'm mulitplying the
note's index by its octave. The reason for multiplying by the octave was
to take into account the vertical part of the note and not just the
horizontal. Looking for a new way to look at both vertical and
horizontal.