Note Duration
In order to determine the duration of a note inside of a MIDI file, the
bpm for each track will have to be calculated. Midi files come in 3
different forms, one where there is only one track, one where multiple
tracks have the same timings and one that has independent timings for
each track. The latter is most common. After calulating the bpm, the
time per tick can be calculated by 60000 / (BPM * PPQ) where 60000 is
the number of milliseconds in a minute.