ClusteringAlgoritm: Distances
Completed a way to get the distance of a MidiFileData object relative to
another. With this will be able to determine closest points to a
centroid. Uncertain how to pick random points for centroid currently.
Probably pick random values that are reasonable for data(i.e bpm within
reasonable bounds, lowest note, etc.)