Clustering Data Entry
Changed the way my clustering algorithm works, now the user must input
the word cluster for the algorithm to run. With each song added, data is
added to a list. Once the user inputs cluster, the program sends all of
the data to the algorithm in an array to do computation on. However,
there are not enough independent  variables to cluster into more
specific groups. Also my randomization of variables might be outside
scope of the data being fed.